mm, compaction: remove unnecessary zone parameter in some instances
A zone parameter is passed into a number of top-level compaction functions
despite the fact that it's already in cache_control.  This is harmless but
it did need an audit to check if zone actually ever changes meaningfully. 
This patches removes the parameter in a number of top-level functions. 
The change could be much deeper but this was enough to briefly clarify the
flow.

No functional change.

Link: http://lkml.kernel.org/r/20190104125011.16071-5-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>