mm, compaction: use the page allocator bulk-free helper for lists of pages
release_pages() is a simpler version of free_unref_page_list() but it
tracks the highest PFN for caching the restart point of the compaction
free scanner.  This patch optionally tracks the highest PFN in the core
helper and converts compaction to use it.  The performance impact is
limited but it should reduce lock contention slightly in some cases.  The
main benefit is removing some partially duplicated code.

Link: http://lkml.kernel.org/r/20190104125011.16071-10-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>