mm, compaction: check early for huge pages encountered by the migration scanner
When scanning for sources or targets, PageCompound is checked for huge
pages as they can be skipped quickly but it happens relatively late after
a lot of setup and checking.  This patch short-cuts the check to make it
earlier.  It might still change when the lock is acquired but this has
less overhead overall.  The free scanner advances but the migration
scanner does not.  Typically the free scanner encounters more movable
blocks that change state over the lifetime of the system and also tends to
scan more aggressively as it's actively filling its portion of the
physical address space with data.  This could change in the future but for
the moment, this worked better in practice and incurred fewer scan
restarts.

The impact on latency and allocation success rates is marginal but the
free scan rates are reduced by 32% and system CPU usage is reduced by
2.6%.  The 2-socket results are not materially different.

Link: http://lkml.kernel.org/r/20190104125011.16071-17-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>