mm, compaction: use free lists to quickly locate a migration source
The migration scanner is a linear scan of a zone with a potentially large
search space.  Furthermore, many pageblocks are unusable such as those
filled with reserved pages or partially filled with pages that cannot
migrate.  These still get scanned in the common case of allocating a THP
and the cost accumulates.

The patch uses a partial search of the free lists to locate a migration
source candidate that is marked as MOVABLE when allocating a THP.  It
prefers picking a block with a larger number of free pages already on the
basis that there are fewer pages to migrate to free the entire block.  The
lowest PFN found during searches is tracked as the basis of the start for
the linear search after the first search of the free list fails.  After
the search, the free list is shuffled so that the next search will not
encounter the same page.  If the search fails then the subsequent searches
will be shorter and the linear scanner is used.

If this search fails, or if the request is for a small or
unmovable/reclaimable allocation then the linear scanner is still used. 
It is somewhat pointless to use the list search in those cases.  Small
free pages must be used for the search and there is no guarantee that
movable pages are located within that block that are contiguous.

                                        4.20.0                 4.20.0
                                failfast-v2r15          findmig-v2r15
Amean     fault-both-1         0.00 (   0.00%)        0.00 *   0.00%*
Amean     fault-both-3      3833.72 (   0.00%)     3505.69 (   8.56%)
Amean     fault-both-5      4967.15 (   0.00%)     5794.13 * -16.65%*
Amean     fault-both-7      7139.19 (   0.00%)     7663.09 (  -7.34%)
Amean     fault-both-12    11326.30 (   0.00%)    10983.36 (   3.03%)
Amean     fault-both-18    16270.70 (   0.00%)    13602.71 *  16.40%*
Amean     fault-both-24    19839.65 (   0.00%)    16145.77 *  18.62%*
Amean     fault-both-30    21707.05 (   0.00%)    19753.82 (   9.00%)
Amean     fault-both-32    21968.16 (   0.00%)    20616.16 (   6.15%)

                                   4.20.0                 4.20.0
                           failfast-v2r15          findmig-v2r15
Percentage huge-1         0.00 (   0.00%)        0.00 (   0.00%)
Percentage huge-3        84.62 (   0.00%)       90.58 (   7.05%)
Percentage huge-5        88.43 (   0.00%)       91.34 (   3.29%)
Percentage huge-7        88.33 (   0.00%)       92.21 (   4.39%)
Percentage huge-12       88.74 (   0.00%)       92.48 (   4.21%)
Percentage huge-18       86.52 (   0.00%)       91.65 (   5.93%)
Percentage huge-24       86.42 (   0.00%)       90.23 (   4.41%)
Percentage huge-30       86.67 (   0.00%)       90.17 (   4.04%)
Percentage huge-32       86.00 (   0.00%)       89.72 (   4.32%)

This shows an improvement in allocation latencies and a slight increase in
allocation success rates.  While not presented, there was a 13% reduction
in migration scanning and a 10% reduction on system CPU usage.  A 2-socket
machine showed similar benefits.

Link: http://lkml.kernel.org/r/20190104125011.16071-12-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>