mm, compaction: do not direct compact remote memory
Remote compaction is expensive and possibly counter-productive.  Locality
is expected to often have better performance characteristics than remote
high-order pages.  For small allocations, it's expected that locality is
generally required or fallbacks are possible.  For larger allocations such
as THP, they are forbidden at the time of writing but if __GFP_THISNODE is
ever removed, then it would still be preferable to fallback to small local
base pages over remote THP in the general case.  kcompactd is still woken
via kswapd so compaction happens eventually.

While this patch potentially has both positive and negative effects, it is
best to avoid the possibility of remote compaction given the cost relative
to any potential benefit.

Link: http://lkml.kernel.org/r/20190104125011.16071-26-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>