mm, compaction: skip pageblocks with reserved pages
Reserved pages are set at boot time, tend to be clustered and almost never
become unreserved.  When isolating pages for either migration sources or
target, skip the entire pageblock is one PageReserved page is encountered
on the grounds that it is highly probable the entire pageblock is
reserved.

The performance impact is relative to the number of reserved pages in the
system and their location so it'll be variable but intuitively it should
make sense.  If the memblock allocator was ever changed to spread reserved
pages throughout the address space then this patch would be impaired but
it would also be considered a bug given that such a change would ruin
fragmentation.

On both 1-socket and 2-socket machines, scan rates are reduced slightly on
workloads that intensively allocate THP while the system is fragmented.

Link: http://lkml.kernel.org/r/20190104125011.16071-7-mgorman@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>