mm, compaction: use free lists to quickly locate a migration target -fix
Full compaction of a node passes in negative orders which can lead to
array boundary issues.  While it could be addressed in the control flow of
the primary loop, it would be fragile so explicitly check for the
condition.  This is a fix for the mmotm patch
broken-out/mm-compaction-use-free-lists-to-quickly-locate-a-migration-target.patch

Link: http://lkml.kernel.org/r/20190109212603.GY31517@techsingularity.net
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Kirill A. Shutemov <kirill@shutemov.name>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>