mm,vmscan: accumulated slab reclaim pressure fixes
Jonathan asked about a divide by zero if a slab with shrinker->seeks==0
comes into the new code path, which appears to be possible because the
shrinker list is locked with a read-write semaphore, that allows multiple
parallel reclaimers.

This led me to take another closer look at the code, and find a
possibility for a small slab with a shrinker->seeks > 4 to not have any
items reclaimed from it at all.

This patch fixes both of these bugs.

Link: http://lkml.kernel.org/r/20190129142831.6a373403@imladris.surriel.com
Signed-off-by: Rik van Riel <riel@surriel.com>
Suggested-by: Jonathan Lemon <bsd@fb.com>
Cc: Chris Mason <clm@fb.com>
Cc: Roman Gushchin <guro@fb.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>