mm/hmm: add helpers for driver to safely take the mmap_sem
The device driver context which holds reference to mirror and thus to core
hmm struct might outlive the mm against which it was created.  To prevent
every driver from having to check for that case, provide a helper that
checks if the mm is still alive and takes the mmap_sem in read mode if so.
If the mm has been destroyed (mmu_notifier release call back did happen)
then we return -EINVAL so that calling code knows that it is trying to do
something against an mm that is no longer valid.

Link: http://lkml.kernel.org/r/20190129165428.3931-11-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Ralph Campbell <rcampbell@nvidia.com>
Cc: John Hubbard <jhubbard@nvidia.com>
Cc: Christian König <christian.koenig@amd.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Felix Kuehling <Felix.Kuehling@amd.com>
Cc: Jason Gunthorpe <jgg@mellanox.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>